# shared/openapi/family-cfo.v1.yaml
openapi: 3.1.0
info:
  title: Family CFO API
  version: 0.1.0
  description: >
    Initial specification for the Family CFO self-hosted backend API.
    This contract is the source of truth for generated Swift and Angular clients.
servers:
  - url: /api/v1
    description: Local Family CFO server
...
paths:
...
  /audit:
    get:
      tags:
        - Audit
      operationId: listAuditEvents
      summary: List the household's audit events
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Audit events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditEventListResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /accounts:
    get:
      tags:
        - Accounts
      operationId: listAccounts
      summary: List accounts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Accounts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountListResponse"
        "401":
          $ref: "#/components/responses/Error"
    post:
      tags:
        - Accounts
      operationId: createAccount
      summary: Create an account
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountCreateRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /accounts/{account_id}:
    patch:
      tags:
        - Accounts
      operationId: updateAccount
      summary: Update an account
      security:
        - bearerAuth: []
      parameters:
        - name: account_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountUpdateRequest"
      responses:
        "200":
          description: Account updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags:
        - Accounts
      operationId: deleteAccount
      summary: Delete an account
      security:
        - bearerAuth: []
      parameters:
        - name: account_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Account deleted
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /accounts/{account_id}/balances:
    post:
      tags:
        - Accounts
      operationId: recordAccountBalance
      summary: Record a new balance for an account
      security:
        - bearerAuth: []
      parameters:
        - name: account_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountBalanceCreateRequest"
      responses:
        "201":
          description: Balance recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
  /bills:
    get:
      tags:
        - Bills
      operationId: listBills
      summary: List recurring bills for the household
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Bills
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillListResponse"
        "401":
          $ref: "#/components/responses/Error"
    post:
      tags:
        - Bills
      operationId: createBill
      summary: Create a recurring bill
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillCreateRequest"
      responses:
        "201":
          description: Bill created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /bills/{bill_id}:
    patch:
      tags:
        - Bills
      operationId: updateBill
      summary: Update a recurring bill
      security:
        - bearerAuth: []
      parameters:
        - name: bill_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillUpdateRequest"
      responses:
        "200":
          description: Bill updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags:
        - Bills
      operationId: deleteBill
      summary: Delete a recurring bill
      security:
        - bearerAuth: []
      parameters:
        - name: bill_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Bill deleted
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    Money:
      type: object
      required:
        - amount_minor
        - currency
      properties:
        amount_minor:
          type: integer
          format: int64
          description: Signed integer amount in minor currency units.
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: USD
...
    RecurringFrequency:
      type: string
      enum:
        - weekly
        - biweekly
        - semimonthly
        - monthly
        - quarterly
        - annual
...
    Bill:
      type: object
      required:
        - id
        - name
        - amount
        - frequency
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        amount:
          $ref: "#/components/schemas/Money"
        frequency:
          $ref: "#/components/schemas/RecurringFrequency"
        next_due_date:
          type: string
          format: date
        account_id:
          type: string
          format: uuid
...
    AccountListResponse:
      type: object
      required:
        - accounts
      properties:
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/Account"
...
    BillListResponse:
      type: object
      required:
        - bills
      properties:
        bills:
          type: array
          items:
            $ref: "#/components/schemas/Bill"
...
    Account:
      type: object
      required:
        - id
        - name
        - type
        - balance
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          $ref: "#/components/schemas/AccountType"
        balance:
          $ref: "#/components/schemas/Money"
    AccountType:
      type: string
      enum:
        - checking
        - savings
        - credit_card
        - brokerage
        - retirement
        - hsa
        - 529
        - mortgage
        - auto_loan
        - student_loan
        - real_estate
        - other_asset
        - other_liability
...
    AccountCreateRequest:
      type: object
      required:
        - name
        - type
        - currency
      properties:
        name:
          type: string
        type:
          $ref: "#/components/schemas/AccountType"
        currency:
          type: string
          minLength: 3
          maxLength: 3
    AccountUpdateRequest:
      type: object
      properties:
        name:
          type: string
        type:
          $ref: "#/components/schemas/AccountType"
    AccountBalanceCreateRequest:
      type: object
      required:
        - balance
      properties:
        balance:
          $ref: "#/components/schemas/Money"
...
    BillCreateRequest:
      type: object
      required:
        - name
        - amount
        - frequency
      properties:
        name:
          type: string
        amount:
          $ref: "#/components/schemas/Money"
        frequency:
          $ref: "#/components/schemas/RecurringFrequency"
        account_id:
          type: string
          format: uuid
        next_due_date:
          type: string
          format: date
    BillUpdateRequest:
      type: object
      properties:
        name:
          type: string
        amount:
          $ref: "#/components/schemas/Money"
        frequency:
          $ref: "#/components/schemas/RecurringFrequency"
        next_due_date:
          type: string
          format: date
...
    AuditEvent:
      type: object
      required:
        - id
        - action
        - entity_type
        - summary
        - created_at
      properties:
        id:
          type: string
          format: uuid
        actor_user_id:
          type: string
          format: uuid
        action:
          type: string
        entity_type:
          type: string
        entity_id:
          type: string
        summary:
          type: string
        created_at:
          type: string
          format: date-time
    AuditEventListResponse:
      type: object
      required:
        - events
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/AuditEvent"
...